+++ src/sentry/api/endpoints/avatar/user.py
@@ -7,7 +7,7 @@ from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import control_silo_endpoint
 from sentry.api.bases.avatar import AvatarMixin
 from sentry.api.bases.user import UserEndpoint
-from sentry.models.avatars.user_avatar import UserAvatar
+from sentry.users.models.user_avatar import UserAvatar
 from sentry.users.services.user.serial import serialize_rpc_user
 from sentry.users.services.user.service import user_service
