+++ src/sentry/models/avatars/__init__.py
@@ -2,12 +2,10 @@ from .base import AvatarBase
 from .control_base import ControlAvatarBase
 from .organization_avatar import OrganizationAvatar
 from .sentry_app_avatar import SentryAppAvatar
-from .user_avatar import UserAvatar
 
 __all__ = (
     "AvatarBase",
     "ControlAvatarBase",
     "OrganizationAvatar",
     "SentryAppAvatar",
-    "UserAvatar",
 )
