+++ src/sentry/web/frontend/user_avatar.py
@@ -1,4 +1,4 @@
-from sentry.models.avatars.user_avatar import UserAvatar
+from sentry.users.models.user_avatar import UserAvatar
 from sentry.web.frontend.base import AvatarPhotoView, control_silo_view
