+++ tests/sentry/models/test_projectownership.py
@@ -1,6 +1,5 @@
 from unittest.mock import patch
 
-from sentry.models.avatars.user_avatar import UserAvatar
 from sentry.models.groupassignee import GroupAssignee
 from sentry.models.groupowner import GroupOwner, GroupOwnerType, OwnerRuleType
 from sentry.models.projectownership import ProjectOwnership
@@ -11,6 +10,7 @@ from sentry.testutils.helpers.datetime import before_now, iso_format
 from sentry.testutils.silo import assume_test_silo_mode_of
 from sentry.testutils.skips import requires_snuba
 from sentry.types.actor import Actor, ActorType
+from sentry.users.models.user_avatar import UserAvatar
 from sentry.users.services.user.service import user_service
 
 pytestmark = requires_snuba
