+++ tests/sentry/users/models/test_user_avatar.py
@@ -1,10 +1,8 @@
 from sentry import options as options_store
-from sentry.models.avatars.organization_avatar import OrganizationAvatar
-from sentry.models.avatars.user_avatar import UserAvatar
 from sentry.models.files.control_file import ControlFile
-from sentry.models.files.file import File
 from sentry.testutils.cases import TestCase
 from sentry.testutils.silo import control_silo_test
+from sentry.users.models.user_avatar import UserAvatar
 
 
 @control_silo_test
@@ -26,17 +24,3 @@ class UserAvatarTestCase(TestCase):
             assert avatar.get_file() is None
             assert UserAvatar.objects.get(id=avatar.id).control_file_id is None
             assert UserAvatar.objects.get(id=avatar.id).get_file() is None
-
-
-class OrganizationAvatarTestCase(TestCase):
-    def test_set_null(self):
-        org = self.create_organization()
-        afile = File.objects.create(name="avatar.png", type=OrganizationAvatar.FILE_TYPE)
-        avatar = OrganizationAvatar.objects.create(organization=org, file_id=afile.id)
-
-        assert avatar.get_file() == afile
-
-        afile.delete()
-        assert avatar.get_file() is None
-        assert OrganizationAvatar.objects.get(id=avatar.id).file_id is None
-        assert OrganizationAvatar.objects.get(id=avatar.id).get_file() is None
